change compliesWith<XXX> method parameter from double to BigDecimal
affected methods: Limit.compliesWith, PlanPhase.compliesWithLimits, Product.compliesWithLimits, Usage.compliesWithLimits

+++ src/main/java/org/killbill/billing/catalog/api/Limit.java
@@ -26,6 +26,6 @@ public interface Limit {
 
     public BigDecimal getMin();
 
-    public boolean compliesWith(double value);
+    public boolean compliesWith(BigDecimal value);
 
 }
